Optimized compilation time on linux
