+++ regamedll/dlls/buttons.cpp
