+++ regamedll/dlls/client.h
@@ -213,7 +213,7 @@ void Custom_Entity_FieldInit(delta_s *pFields);
 void Custom_Encode(struct delta_s *pFields, const unsigned char *from, const unsigned char *to);
 void RegisterEncoders(void);
 int GetWeaponData(edict_s *player, struct weapon_data_s *info);
-void UpdateClientData(const edict_s *ent, int sendweapons, clientdata_s *cd);
+void UpdateClientData(const edict_s *ent, int sendweapons, struct clientdata_s *cd);
 void CmdStart(const edict_t *player, const struct usercmd_s *cmd, unsigned int random_seed);
 void CmdEnd(const edict_t *player);
 int ConnectionlessPacket(const struct netadr_s *net_from, const char *args, char *response_buffer, int *response_buffer_size);
