+++ regamedll/dlls/h_ai.cpp
@@ -1,8 +1,5 @@
 #include "precompiled.h"
 
-#define NUM_LATERAL_CHECKS		13	// how many checks are made on each side of a monster looking for lateral cover
-#define NUM_LATERAL_LOS_CHECKS		6	// how many checks are made on each side of a monster looking for lateral cover
-
 /*
 * Globals initialization
 */
