+++ regamedll/dlls/tutor_cs_tutor.cpp
@@ -1,5 +1,4 @@
 #include "precompiled.h"
-#include <string>
 
 /*
 * Globals initialization
