+++ regamedll/dlls/world.cpp
@@ -1,9 +1,5 @@
 #include "precompiled.h"
 
-#define SF_WORLD_DARK		0x0001		// Fade from black at startup
-#define SF_WORLD_TITLE		0x0002		// Display game title at startup
-#define SF_WORLD_FORCETEAM	0x0004		// Force teams
-
 edict_t *g_pBodyQueueHead;
 CGlobalState gGlobalState;
 float g_flWeaponCheat;
