+++ regamedll/dlls/wpn_shared.h
@@ -1,0 +1,598 @@
+#pragma once
+
+//AK47
+#define AK47_MAX_SPEED			221
+#define AK47_DAMAGE			36
+#define AK47_RANGE_MODIFER		0.98
+#define AK47_RELOAD_TIME		2.45
+
+enum ak47_e
+{
+	AK47_IDLE1,
+	AK47_RELOAD,
+	AK47_DRAW,
+	AK47_SHOOT1,
+	AK47_SHOOT2,
+	AK47_SHOOT3
+};
+
+
+
+//AUG
+#define AUG_MAX_SPEED			240
+#define AUG_DAMAGE			32
+#define AUG_RANGE_MODIFER		0.96
+#define AUG_RELOAD_TIME			3.3
+
+enum aug_e
+{
+	AUG_IDLE1,
+	AUG_RELOAD,
+	AUG_DRAW,
+	AUG_SHOOT1,
+	AUG_SHOOT2,
+	AUG_SHOOT3
+};
+
+
+
+//AWP
+#define AWP_MAX_SPEED		210
+#define AWP_MAX_SPEED_ZOOM	150
+#define AWP_DAMAGE		115
+#define AWP_RANGE_MODIFER	0.99
+#define AWP_RELOAD_TIME		2.5
+
+enum awp_e
+{
+	AWP_IDLE,
+	AWP_SHOOT,
+	AWP_SHOOT2,
+	AWP_SHOOT3,
+	AWP_RELOAD,
+	AWP_DRAW,
+};
+
+
+
+//C4
+#define C4_MAX_AMMO		1
+#define C4_MAX_SPEED		250.0
+#define C4_ARMING_ON_TIME	3.0
+
+enum c4_e
+{
+	C4_IDLE1,
+	C4_DRAW,
+	C4_DROP,
+	C4_ARM
+};
+
+
+
+//Deagle
+#define DEAGLE_MAX_SPEED	250
+#define DEAGLE_DAMAGE		54
+#define DEAGLE_RANGE_MODIFER	0.81
+#define DEAGLE_RELOAD_TIME	2.2
+
+enum deagle_e
+{
+	DEAGLE_IDLE1,
+	DEAGLE_SHOOT1,
+	DEAGLE_SHOOT2,
+	DEAGLE_SHOOT_EMPTY,
+	DEAGLE_RELOAD,
+	DEAGLE_DRAW
+};
+
+
+
+//Elites
+#define ELITE_MAX_SPEED		250
+#define ELITE_RELOAD_TIME	4.5
+#define ELITE_DAMAGE		36
+#define ELITE_RANGE_MODIFER	0.75
+
+enum elite_e
+{
+	ELITE_IDLE,
+	ELITE_IDLE_LEFTEMPTY,
+	ELITE_SHOOTLEFT1,
+	ELITE_SHOOTLEFT2,
+	ELITE_SHOOTLEFT3,
+	ELITE_SHOOTLEFT4,
+	ELITE_SHOOTLEFT5,
+	ELITE_SHOOTLEFTLAST,
+	ELITE_SHOOTRIGHT1,
+	ELITE_SHOOTRIGHT2,
+	ELITE_SHOOTRIGHT3,
+	ELITE_SHOOTRIGHT4,
+	ELITE_SHOOTRIGHT5,
+	ELITE_SHOOTRIGHTLAST,
+	ELITE_RELOAD,
+	ELITE_DRAW
+};
+
+
+
+//Famas
+#define FAMAS_MAX_SPEED		240
+#define FAMAS_RELOAD_TIME	3.3
+#define FAMAS_DAMAGE		30
+#define FAMAS_DAMAGE_BURST	34
+#define FAMAS_RANGE_MODIFER	0.96
+
+enum famas_e
+{
+	FAMAS_IDLE1,
+	FAMAS_RELOAD,
+	FAMAS_DRAW,
+	FAMAS_SHOOT1,
+	FAMAS_SHOOT2,
+	FAMAS_SHOOT3
+};
+
+
+
+//Fiveseven
+#define FIVESEVEN_MAX_SPEED		250
+#define FIVESEVEN_DAMAGE		20
+#define FIVESEVEN_RANGE_MODIFER		0.885
+#define FIVESEVEN_RELOAD_TIME		2.7
+
+enum fiveseven_e
+{
+	FIVESEVEN_IDLE,
+	FIVESEVEN_SHOOT1,
+	FIVESEVEN_SHOOT2,
+	FIVESEVEN_SHOOT_EMPTY,
+	FIVESEVEN_RELOAD,
+	FIVESEVEN_DRAW
+};
+
+
+
+//Flashbang
+#define FLASHBANG_MAX_SPEED		250
+#define FLASHBANG_MAX_SPEED_SHIELD	180
+
+enum flashbang_e
+{
+	FLASHBANG_IDLE,
+	FLASHBANG_PULLPIN,
+	FLASHBANG_THROW,
+	FLASHBANG_DRAW
+};
+
+
+
+//g3sg1
+#define G3SG1_MAX_SPEED		210
+#define G3SG1_MAX_SPEED_ZOOM	150
+#define G3SG1_DAMAGE		80
+#define G3SG1_RANGE_MODIFER	0.98
+#define G3SG1_RELOAD_TIME	3.5
+
+enum g3sg1_e
+{
+	G3SG1_IDLE,
+	G3SG1_SHOOT,
+	G3SG1_SHOOT2,
+	G3SG1_RELOAD,
+	G3SG1_DRAW
+};
+
+
+
+//galil
+#define GALIL_MAX_SPEED			240
+#define GALIL_DAMAGE			30
+#define GALIL_RANGE_MODIFER		0.98
+#define GALIL_RELOAD_TIME		2.45
+
+enum galil_e
+{
+	GALIL_IDLE1,
+	GALIL_RELOAD,
+	GALIL_DRAW,
+	GALIL_SHOOT1,
+	GALIL_SHOOT2,
+	GALIL_SHOOT3
+};
+
+
+
+//glock18
+#define GLOCK18_MAX_SPEED		250
+#define GLOCK18_DAMAGE			25
+#define GLOCK18_RANGE_MODIFER		0.75
+#define GLOCK18_RELOAD_TIME		2.2
+
+enum glock18_e
+{
+	GLOCK18_IDLE1,
+	GLOCK18_IDLE2,
+	GLOCK18_IDLE3,
+	GLOCK18_SHOOT,
+	GLOCK18_SHOOT2,
+	GLOCK18_SHOOT3,
+	GLOCK18_SHOOT_EMPTY,
+	GLOCK18_RELOAD,
+	GLOCK18_DRAW,
+	GLOCK18_HOLSTER,
+	GLOCK18_ADD_SILENCER,
+	GLOCK18_DRAW2,
+	GLOCK18_RELOAD2
+};
+
+enum glock18_shield_e
+{
+	GLOCK18_SHIELD_IDLE1,
+	GLOCK18_SHIELD_SHOOT,
+	GLOCK18_SHIELD_SHOOT2,
+	GLOCK18_SHIELD_SHOOT_EMPTY,
+	GLOCK18_SHIELD_RELOAD,
+	GLOCK18_SHIELD_DRAW,
+	GLOCK18_SHIELD_IDLE,
+	GLOCK18_SHIELD_UP,
+	GLOCK18_SHIELD_DOWN
+};
+
+
+
+//hegrenade
+#define HEGRENADE_MAX_SPEED		250
+#define HEGRENADE_MAX_SPEED_SHIELD	180
+
+enum hegrenade_e
+{
+	HEGRENADE_IDLE,
+	HEGRENADE_PULLPIN,
+	HEGRENADE_THROW,
+	HEGRENADE_DRAW
+};
+
+
+
+//knife
+#define KNIFE_BODYHIT_VOLUME		128
+#define KNIFE_WALLHIT_VOLUME		512
+#define KNIFE_MAX_SPEED			250
+#define KNIFE_MAX_SPEED_SHIELD		180
+
+enum knife_e
+{
+	KNIFE_IDLE,
+	KNIFE_ATTACK1HIT,
+	KNIFE_ATTACK2HIT,
+	KNIFE_DRAW,
+	KNIFE_STABHIT,
+	KNIFE_STABMISS,
+	KNIFE_MIDATTACK1HIT,
+	KNIFE_MIDATTACK2HIT
+};
+
+enum knife_shield_e
+{
+	KNIFE_SHIELD_IDLE,
+	KNIFE_SHIELD_SLASH,
+	KNIFE_SHIELD_ATTACKHIT,
+	KNIFE_SHIELD_DRAW,
+	KNIFE_SHIELD_UPIDLE,
+	KNIFE_SHIELD_UP,
+	KNIFE_SHIELD_DOWN
+};
+
+
+
+//m3
+#define M3_MAX_SPEED		230
+#define M3_CONE_VECTOR		Vector(0.0675, 0.0675, 0.0)	// special shotgun spreads
+
+enum m3_e
+{
+	M3_IDLE,
+	M3_FIRE1,
+	M3_FIRE2,
+	M3_RELOAD,
+	M3_PUMP,
+	M3_START_RELOAD,
+	M3_DRAW,
+	M3_HOLSTER
+};
+
+
+
+//m4a1
+#define M4A1_MAX_SPEED		230
+#define M4A1_DAMAGE		32
+#define M4A1_DAMAGE_SIL		33
+#define M4A1_RANGE_MODIFER      0.97
+#define M4A1_RANGE_MODIFER_SIL  0.95
+#define M4A1_RELOAD_TIME	3.05
+
+enum m4a1_e
+{
+	M4A1_IDLE,
+	M4A1_SHOOT1,
+	M4A1_SHOOT2,
+	M4A1_SHOOT3,
+	M4A1_RELOAD,
+	M4A1_DRAW,
+	M4A1_ATTACH_SILENCER,
+	M4A1_UNSIL_IDLE,
+	M4A1_UNSIL_SHOOT1,
+	M4A1_UNSIL_SHOOT2,
+	M4A1_UNSIL_SHOOT3,
+	M4A1_UNSIL_RELOAD,
+	M4A1_UNSIL_DRAW,
+	M4A1_DETACH_SILENCER
+};
+
+
+
+//m249
+#define M249_MAX_SPEED			220
+#define M249_DAMAGE			32
+#define M249_RANGE_MODIFER		0.97
+#define M249_RELOAD_TIME		4.7
+
+enum m249_e
+{
+	M249_IDLE1,
+	M249_SHOOT1,
+	M249_SHOOT2,
+	M249_RELOAD,
+	M249_DRAW
+};
+
+
+
+//mac10
+#define MAC10_MAX_SPEED			250
+#define MAC10_DAMAGE			29
+#define MAC10_RANGE_MODIFER		0.82
+#define MAC10_RELOAD_TIME		3.15
+
+enum mac10_e
+{
+	MAC10_IDLE1,
+	MAC10_RELOAD,
+	MAC10_DRAW,
+	MAC10_SHOOT1,
+	MAC10_SHOOT2,
+	MAC10_SHOOT3
+};
+
+
+
+//mp5navy
+#define MP5N_MAX_SPEED			250
+#define MP5N_DAMAGE			26
+#define MP5N_RANGE_MODIFER		0.84
+#define MP5N_RELOAD_TIME		2.63
+
+enum mp5n_e
+{
+	MP5N_IDLE1,
+	MP5N_RELOAD,
+	MP5N_DRAW,
+	MP5N_SHOOT1,
+	MP5N_SHOOT2,
+	MP5N_SHOOT3
+};
+
+
+
+//p90
+#define P90_MAX_SPEED		245
+#define P90_DAMAGE		21
+#define P90_RANGE_MODIFER	0.885
+#define P90_RELOAD_TIME		3.4
+
+enum p90_e
+{
+	P90_IDLE1,
+	P90_RELOAD,
+	P90_DRAW,
+	P90_SHOOT1,
+	P90_SHOOT2,
+	P90_SHOOT3
+};
+
+
+
+//p228
+#define P228_MAX_SPEED		250
+#define P228_DAMAGE		32
+#define P228_RANGE_MODIFER	0.8
+#define P228_RELOAD_TIME	2.7
+
+enum p228_e
+{
+	P228_IDLE,
+	P228_SHOOT1,
+	P228_SHOOT2,
+	P228_SHOOT3,
+	P228_SHOOT_EMPTY,
+	P228_RELOAD,
+	P228_DRAW
+};
+
+enum p228_shield_e
+{
+	P228_SHIELD_IDLE,
+	P228_SHIELD_SHOOT1,
+	P228_SHIELD_SHOOT2,
+	P228_SHIELD_SHOOT_EMPTY,
+	P228_SHIELD_RELOAD,
+	P228_SHIELD_DRAW,
+	P228_SHIELD_IDLE_UP,
+	P228_SHIELD_UP,
+	P228_SHIELD_DOWN
+};
+
+
+
+//scout
+#define SCOUT_MAX_SPEED			260
+#define SCOUT_MAX_SPEED_ZOOM		220
+#define SCOUT_DAMAGE			75
+#define SCOUT_RANGE_MODIFER		0.98
+#define SCOUT_RELOAD_TIME		2
+
+enum scout_e
+{
+	SCOUT_IDLE,
+	SCOUT_SHOOT,
+	SCOUT_SHOOT2,
+	SCOUT_RELOAD,
+	SCOUT_DRAW
+};
+
+
+
+//sg550
+#define SG550_MAX_SPEED			210
+#define SG550_MAX_SPEED_ZOOM		150
+#define SG550_DAMAGE			70
+#define SG550_RANGE_MODIFER		0.98
+#define SG550_RELOAD_TIME		3.35
+
+enum sg550_e
+{
+	SG550_IDLE,
+	SG550_SHOOT,
+	SG550_SHOOT2,
+	SG550_RELOAD,
+	SG550_DRAW
+};
+
+
+
+//sg552
+#define SG552_MAX_SPEED			235
+#define SG552_MAX_SPEED_ZOOM		200
+#define SG552_DAMAGE			33
+#define SG552_RANGE_MODIFER		0.955
+#define SG552_RELOAD_TIME		3
+
+enum sg552_e
+{
+	SG552_IDLE1,
+	SG552_RELOAD,
+	SG552_DRAW,
+	SG552_SHOOT1,
+	SG552_SHOOT2,
+	SG552_SHOOT3
+};
+
+
+
+//smokegrenade
+#define SMOKEGRENADE_MAX_SPEED		250
+#define SMOKEGRENADE_MAX_SPEED_SHIELD	180
+
+enum smokegrenade_e
+{
+	SMOKEGRENADE_IDLE,
+	SMOKEGRENADE_PINPULL,
+	SMOKEGRENADE_THROW,
+	SMOKEGRENADE_DRAW
+};
+
+
+
+//tmp
+#define TMP_MAX_SPEED			250
+#define TMP_DAMAGE			20
+#define TMP_RANGE_MODIFER		0.85
+#define TMP_RELOAD_TIME			2.12
+
+enum tmp_e
+{
+	TMP_IDLE1,
+	TMP_RELOAD,
+	TMP_DRAW,
+	TMP_SHOOT1,
+	TMP_SHOOT2,
+	TMP_SHOOT3
+};
+
+
+
+//ump45
+#define UMP45_MAX_SPEED			250
+#define UMP45_DAMAGE			30
+#define UMP45_RANGE_MODIFER		0.82
+#define UMP45_RELOAD_TIME		3.5
+
+enum ump45_e
+{
+	UMP45_IDLE1,
+	UMP45_RELOAD,
+	UMP45_DRAW,
+	UMP45_SHOOT1,
+	UMP45_SHOOT2,
+	UMP45_SHOOT3
+};
+
+
+
+//tmp
+#define USP_MAX_SPEED		250
+#define USP_DAMAGE		34
+#define USP_DAMAGE_SIL		30
+#define USP_RANGE_MODIFER	0.79
+#define USP_RELOAD_TIME		2.7
+
+enum usp_e
+{
+	USP_IDLE,
+	USP_SHOOT1,
+	USP_SHOOT2,
+	USP_SHOOT3,
+	USP_SHOOT_EMPTY,
+	USP_RELOAD,
+	USP_DRAW,
+	USP_ATTACH_SILENCER,
+	USP_UNSIL_IDLE,
+	USP_UNSIL_SHOOT1,
+	USP_UNSIL_SHOOT2,
+	USP_UNSIL_SHOOT3,
+	USP_UNSIL_SHOOT_EMPTY,
+	USP_UNSIL_RELOAD,
+	USP_UNSIL_DRAW,
+	USP_DETACH_SILENCER
+};
+
+enum usp_shield_e
+{
+	USP_SHIELD_IDLE,
+	USP_SHIELD_SHOOT1,
+	USP_SHIELD_SHOOT2,
+	USP_SHIELD_SHOOT_EMPTY,
+	USP_SHIELD_RELOAD,
+	USP_SHIELD_DRAW,
+	USP_SHIELD_UP_IDLE,
+	USP_SHIELD_UP,
+	USP_SHIELD_DOWN
+};
+
+
+
+//xm1014
+#define XM1014_MAX_SPEED	240
+#define XM1014_CONE_VECTOR	Vector(0.0725, 0.0725, 0.0)	// special shotgun spreads
+
+enum xm1014_e
+{
+	XM1014_IDLE,
+	XM1014_FIRE1,
+	XM1014_FIRE2,
+	XM1014_RELOAD,
+	XM1014_PUMP,
+	XM1014_START_RELOAD,
+	XM1014_DRAW
+};
