+++ regamedll/dlls/wpn_shared/wpn_elite.cpp
@@ -1,31 +1,5 @@
 #include "precompiled.h"
 
-#define ELITE_MAX_SPEED		250
-#define ELITE_RELOAD_TIME	4.5
-
-#define ELITE_DAMAGE		36
-#define ELITE_RANGE_MODIFER	0.75
-
-enum elite_e
-{
-	ELITE_IDLE,
-	ELITE_IDLE_LEFTEMPTY,
-	ELITE_SHOOTLEFT1,
-	ELITE_SHOOTLEFT2,
-	ELITE_SHOOTLEFT3,
-	ELITE_SHOOTLEFT4,
-	ELITE_SHOOTLEFT5,
-	ELITE_SHOOTLEFTLAST,
-	ELITE_SHOOTRIGHT1,
-	ELITE_SHOOTRIGHT2,
-	ELITE_SHOOTRIGHT3,
-	ELITE_SHOOTRIGHT4,
-	ELITE_SHOOTRIGHT5,
-	ELITE_SHOOTRIGHTLAST,
-	ELITE_RELOAD,
-	ELITE_DRAW
-};
-
 /* <25137f> ../cstrike/dlls/wpn_shared/wpn_elite.cpp:52 */
 LINK_ENTITY_TO_CLASS(weapon_elite, CELITE);
