+++ regamedll/dlls/wpn_shared/wpn_glock18.cpp
@@ -1,42 +1,5 @@
 #include "precompiled.h"
 
-#define GLOCK18_MAX_SPEED		250
-
-#define GLOCK18_DAMAGE			25
-#define GLOCK18_RANGE_MODIFER		0.75
-
-#define GLOCK18_RELOAD_TIME		2.2
-
-enum glock18_e
-{
-	GLOCK18_IDLE1,
-	GLOCK18_IDLE2,
-	GLOCK18_IDLE3,
-	GLOCK18_SHOOT,
-	GLOCK18_SHOOT2,
-	GLOCK18_SHOOT3,
-	GLOCK18_SHOOT_EMPTY,
-	GLOCK18_RELOAD,
-	GLOCK18_DRAW,
-	GLOCK18_HOLSTER,
-	GLOCK18_ADD_SILENCER,
-	GLOCK18_DRAW2,
-	GLOCK18_RELOAD2
-};
-
-enum glock18_shield_e
-{
-	GLOCK18_SHIELD_IDLE1,
-	GLOCK18_SHIELD_SHOOT,
-	GLOCK18_SHIELD_SHOOT2,
-	GLOCK18_SHIELD_SHOOT_EMPTY,
-	GLOCK18_SHIELD_RELOAD,
-	GLOCK18_SHIELD_DRAW,
-	GLOCK18_SHIELD_IDLE,
-	GLOCK18_SHIELD_UP,
-	GLOCK18_SHIELD_DOWN
-};
-
 /* <2662bb> ../cstrike/dlls/wpn_shared/wpn_glock18.cpp:68 */
 LINK_ENTITY_TO_CLASS(weapon_glock18, CGLOCK18);
