+++ regamedll/dlls/wpn_shared/wpn_knife.cpp
@@ -1,34 +1,5 @@
 #include "precompiled.h"
 
-#define KNIFE_BODYHIT_VOLUME		128
-#define KNIFE_WALLHIT_VOLUME		512
-
-#define KNIFE_MAX_SPEED			250
-#define KNIFE_MAX_SPEED_SHIELD		180
-
-enum knife_e
-{
-	KNIFE_IDLE,
-	KNIFE_ATTACK1HIT,
-	KNIFE_ATTACK2HIT,
-	KNIFE_DRAW,
-	KNIFE_STABHIT,
-	KNIFE_STABMISS,
-	KNIFE_MIDATTACK1HIT,
-	KNIFE_MIDATTACK2HIT
-};
-
-enum knife_shield_e
-{
-	KNIFE_SHIELD_IDLE,
-	KNIFE_SHIELD_SLASH,
-	KNIFE_SHIELD_ATTACKHIT,
-	KNIFE_SHIELD_DRAW,
-	KNIFE_SHIELD_UPIDLE,
-	KNIFE_SHIELD_UP,
-	KNIFE_SHIELD_DOWN
-};
-
 /* <270ca6> ../cstrike/dlls/wpn_shared/wpn_knife.cpp:28 */
 LINK_ENTITY_TO_CLASS(weapon_knife, CKnife);
