+++ regamedll/dlls/wpn_shared/wpn_m4a1.cpp
@@ -1,33 +1,5 @@
 #include "precompiled.h"
 
-#define M4A1_MAX_SPEED		230
-
-#define M4A1_DAMAGE		32
-#define M4A1_DAMAGE_SIL		33
-
-#define M4A1_RANGE_MODIFER      0.97
-#define M4A1_RANGE_MODIFER_SIL  0.95
-
-#define M4A1_RELOAD_TIME	3.05
-
-enum m4a1_e
-{
-	M4A1_IDLE,
-	M4A1_SHOOT1,
-	M4A1_SHOOT2,
-	M4A1_SHOOT3,
-	M4A1_RELOAD,
-	M4A1_DRAW,
-	M4A1_ATTACH_SILENCER,
-	M4A1_UNSIL_IDLE,
-	M4A1_UNSIL_SHOOT1,
-	M4A1_UNSIL_SHOOT2,
-	M4A1_UNSIL_SHOOT3,
-	M4A1_UNSIL_RELOAD,
-	M4A1_UNSIL_DRAW,
-	M4A1_DETACH_SILENCER
-};
-
 /* <281342> ../cstrike/dlls/wpn_shared/wpn_m4a1.cpp:60 */
 LINK_ENTITY_TO_CLASS(weapon_m4a1, CM4A1);
