+++ regamedll/dlls/wpn_shared/wpn_p228.cpp
@@ -1,36 +1,5 @@
 #include "precompiled.h"
 
-#define P228_MAX_SPEED		250
-
-#define P228_DAMAGE		32
-#define P228_RANGE_MODIFER	0.8
-
-#define P228_RELOAD_TIME	2.7
-
-enum p228_e
-{
-	P228_IDLE,
-	P228_SHOOT1,
-	P228_SHOOT2,
-	P228_SHOOT3,
-	P228_SHOOT_EMPTY,
-	P228_RELOAD,
-	P228_DRAW
-};
-
-enum p228_shield_e
-{
-	P228_SHIELD_IDLE,
-	P228_SHIELD_SHOOT1,
-	P228_SHIELD_SHOOT2,
-	P228_SHIELD_SHOOT_EMPTY,
-	P228_SHIELD_RELOAD,
-	P228_SHIELD_DRAW,
-	P228_SHIELD_IDLE_UP,
-	P228_SHIELD_UP,
-	P228_SHIELD_DOWN
-};
-
 /* <291149> ../cstrike/dlls/wpn_shared/wpn_p228.cpp:63 */
 LINK_ENTITY_TO_CLASS(weapon_p228, CP228);
