+++ regamedll/dlls/wpn_shared/wpn_usp.cpp
@@ -1,47 +1,5 @@
 #include "precompiled.h"
 
-#define USP_MAX_SPEED		250
-
-#define USP_DAMAGE		34
-#define USP_DAMAGE_SIL		30
-
-#define USP_RANGE_MODIFER	0.79
-
-#define USP_RELOAD_TIME		2.7
-
-enum usp_e
-{
-	USP_IDLE,
-	USP_SHOOT1,
-	USP_SHOOT2,
-	USP_SHOOT3,
-	USP_SHOOT_EMPTY,
-	USP_RELOAD,
-	USP_DRAW,
-	USP_ATTACH_SILENCER,
-	USP_UNSIL_IDLE,
-	USP_UNSIL_SHOOT1,
-	USP_UNSIL_SHOOT2,
-	USP_UNSIL_SHOOT3,
-	USP_UNSIL_SHOOT_EMPTY,
-	USP_UNSIL_RELOAD,
-	USP_UNSIL_DRAW,
-	USP_DETACH_SILENCER
-};
-
-enum usp_shield_e
-{
-	USP_SHIELD_IDLE,
-	USP_SHIELD_SHOOT1,
-	USP_SHIELD_SHOOT2,
-	USP_SHIELD_SHOOT_EMPTY,
-	USP_SHIELD_RELOAD,
-	USP_SHIELD_DRAW,
-	USP_SHIELD_UP_IDLE,
-	USP_SHIELD_UP,
-	USP_SHIELD_DOWN
-};
-
 /* <2bb045> ../cstrike/dlls/wpn_shared/wpn_usp.cpp:68 */
 LINK_ENTITY_TO_CLASS(weapon_usp, CUSP);
