+++ regamedll/engine/unicode_strtools.cpp
