+++ regamedll/game_shared/perf_counter.h
